Junhyuk feat

See merge request cloud_track/class_01/web_project3/team01/thisthat-be!15

+++ src/main/java/com/example/demo/module/review/service/ReviewServiceImpl.java
@@ -2,8 +2,11 @@ package com.example.demo.module.review.service;
 
 import com.example.demo.module.common.result.PageResult;
 import com.example.demo.module.hashtag.repository.HashtagRepository;
-import com.example.demo.module.image.repository.ReviewImageRepository;
+import com.example.demo.module.image.entity.ReviewImage;
+import com.example.demo.module.image.service.ImageService;
 import com.example.demo.module.medicine.repository.MedicineRepository;
+import com.example.demo.module.point.entity.PointHistory;
+import com.example.demo.module.point.repository.PointHistoryRepository;
 import com.example.demo.module.review.dto.payload.ReviewEditPayload;
 import com.example.demo.module.review.dto.payload.ReviewPayload;
 import com.example.demo.module.review.dto.result.ReviewResult;
@@ -11,14 +14,21 @@ import com.example.demo.module.review.entity.Review;
 import com.example.demo.module.review.entity.ReviewHashtag;
 import com.example.demo.module.review.repository.ReviewHashtagRepository;
 import com.example.demo.module.review.repository.ReviewRepository;
+import com.example.demo.module.user.entity.User;
+import com.example.demo.module.user.repository.UserRepository;
 import com.example.demo.util.mapper.ReviewMapper;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.IOException;
+import java.util.List;
 import java.util.NoSuchElementException;
+
+import static com.example.demo.module.point.entity.ReserveUse.RESERVE;
 
 @Service
 @RequiredArgsConstructor
@@ -27,15 +37,24 @@ public class ReviewServiceImpl implements ReviewService {
     private final ReviewRepository reviewRepository;
     private final MedicineRepository medicineRepository;
     private final ReviewMapper reviewMapper;
-    private final ReviewImageRepository imageRepository;
     private final ReviewHashtagRepository reviewHashtagRepository;
     private final HashtagRepository hashtagRepository;
+    private final UserRepository userRepository;
+    private final PointHistoryRepository pointHistoryRepository;
+
+    @Value("${point.review}")
+    private Integer reviewCreatePoint;
 
     @Override
     @Transactional
-    public Long save(ReviewPayload reviewPayload) {
+    public Long save(Long userId, ReviewPayload reviewPayload) {
 
         //TODO : 이미지 저장 로직 추가 필요
+        User user = userRepository.findById(userId).orElseThrow(() -> new NoSuchElementException("해당하는 유저가 없습니다."));
+
+        if (reviewRepository.existsByMedicineIdAndUserUserId(reviewPayload.getMedicineId(), user.getUserId())) {
+            throw new IllegalArgumentException("이미 후기를 작성한 영양제 입니다.");
+        }
 
         Review review = reviewRepository.save(
                 Review.builder()
@@ -44,6 +63,7 @@ public class ReviewServiceImpl implements ReviewService {
                         .star(reviewPayload.getStar())
                         .heartCount(0)
                         .medicine(medicineRepository.findById(reviewPayload.getMedicineId()).orElseThrow(() -> new NoSuchElementException("해당하는 영양제가 없습니다.")))
+                        .user(user)
                         .build());
 
         reviewPayload.getTagList().forEach(
@@ -52,6 +72,13 @@ public class ReviewServiceImpl implements ReviewService {
                                 .review(review)
                                 .hashtag(hashtagRepository.findById(ht).orElseThrow())
                                 .build()));
+        pointHistoryRepository.save(
+                PointHistory.builder()
+                        .domain("review")
+                        .user(user)
+                        .changedValue(reviewCreatePoint)
+                        .pointSum(user.reviewPoint(reviewCreatePoint))
+                        .reserveUse(RESERVE).build());
 
         return review.getId();
     }
@@ -65,8 +92,11 @@ public class ReviewServiceImpl implements ReviewService {
     //이미지 수정은 따로 뺼 예정
     @Override
     @Transactional
-    public Long editReview(Long reviewId, ReviewEditPayload reviewEditPayload) {
-
+    public Long editReview(Long userId, Long reviewId, ReviewEditPayload reviewEditPayload) {
+        if (!reviewRepository.findById(reviewId).orElseThrow(() -> new NoSuchElementException("후기를 찾을 수 없습니다."))
+                .getUser().getUserId().equals(userId)) {
+            throw new IllegalArgumentException("후기 작성자만 수정할 수 있습니다.");
+        }
         //없어진 해쉬태그 삭제
         reviewHashtagRepository.findAllByReviewId(reviewId).forEach(ht -> {
             if (!reviewEditPayload.getTagList().contains(ht)) {
@@ -89,8 +119,13 @@ public class ReviewServiceImpl implements ReviewService {
         return review.update(reviewEditPayload);
     }
 
+    @Transactional
     @Override
-    public Long deleteByReviewId(Long reviewId) {
+    public Long deleteByReviewId(Long userId, Long reviewId) {
+        if (!reviewRepository.findById(reviewId).orElseThrow(() -> new NoSuchElementException("후기를 찾을 수 없습니다."))
+                .getUser().getUserId().equals(userId)) {
+            throw new IllegalArgumentException("후기 작성자만 삭제할 수 있습니다.");
+        }
         reviewRepository.deleteById(reviewId);
         return reviewId;
     }
